requirements.txt
beautifulsoup4==4.11.1
blinker==1.4
cachetools==5.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
cssmin==0.2.0
Deprecated==1.2.13
docutils==0.18.1
feedgenerator==2.0.0
frozendict==2.3.2
ghp-import==2.1.0
idna==3.3
invoke==1.7.1
isodate==0.6.1
Jinja2==3.1.2
laconia==0.6.0
livereload==2.6.3
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
pelican==4.7.2
pelican-sitemap==1.0.2
pelican-webassets==2.0.0
pycparser==2.21
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.4.0
PyLD==2.0.3
PyNaCl==1.5.0
pyparsing==3.0.9
pyScss==1.4.0
python-dateutil==2.8.2
pytz==2022.1
rdflib==6.1.1
requests==2.27.1
rich==12.4.1
sh==1.14.2
six==1.16.0
smartypants==2.0.1
soupsieve==2.3.2.post1
tornado==6.1
typogrify==2.0.7
Unidecode==1.3.4
urllib3==1.26.9
voltaire==0.11.1
webassets==2.0
wrapt==1.14.1
